rag/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
redis==5.0.1
asyncpg==0.29.0
langchain==0.2.11
langchain-openai==0.1.20
sentence-transformers==2.2.2
faiss-cpu==1.7.4
